+++ py/circuitpy_mpconfig.mk
@@ -61,6 +61,11 @@ CIRCUITPY_AUDIOIO = $(CIRCUITPY_FULL_BUILD)
 endif
 CFLAGS += -DCIRCUITPY_AUDIOIO=$(CIRCUITPY_AUDIOIO)
 
+ifndef CIRCUITPY_AUDIOCORE
+CIRCUITPY_AUDIOCORE = $(CIRCUITPY_AUDIOIO)
+endif
+CFLAGS += -DCIRCUITPY_AUDIOCORE=$(CIRCUITPY_AUDIOCORE)
+
 ifndef CIRCUITPY_BITBANGIO
 CIRCUITPY_BITBANGIO = $(CIRCUITPY_FULL_BUILD)
 endif
