+++ shared-bindings/audiocore/RawSample.h
@@ -24,23 +24,21 @@
  * THE SOFTWARE.
  */
 
-#ifndef MICROPY_INCLUDED_SHARED_BINDINGS_AUDIOIO_WAVEFILE_H
-#define MICROPY_INCLUDED_SHARED_BINDINGS_AUDIOIO_WAVEFILE_H
+#ifndef MICROPY_INCLUDED_SHARED_BINDINGS_AUDIOIO_RAWSAMPLE_H
+#define MICROPY_INCLUDED_SHARED_BINDINGS_AUDIOIO_RAWSAMPLE_H
 
-#include "common-hal/audioio/AudioOut.h"
 #include "common-hal/microcontroller/Pin.h"
-#include "extmod/vfs_fat.h"
+#include "shared-module/audiocore/RawSample.h"
 
-extern const mp_obj_type_t audioio_wavefile_type;
+extern const mp_obj_type_t audioio_rawsample_type;
 
-void common_hal_audioio_wavefile_construct(audioio_wavefile_obj_t* self,
-    pyb_file_obj_t* file);
+void common_hal_audioio_rawsample_construct(audioio_rawsample_obj_t* self,
+    uint8_t* buffer, uint32_t len, uint8_t bytes_per_sample, bool samples_signed,
+    uint8_t channel_count, uint32_t sample_rate);
 
-void common_hal_audioio_wavefile_deinit(audioio_wavefile_obj_t* self);
-bool common_hal_audioio_wavefile_deinited(audioio_wavefile_obj_t* self);
-uint32_t common_hal_audioio_wavefile_get_sample_rate(audioio_wavefile_obj_t* self);
-void common_hal_audioio_wavefile_set_sample_rate(audioio_wavefile_obj_t* self, uint32_t sample_rate);
-uint8_t common_hal_audioio_wavefile_get_bits_per_sample(audioio_wavefile_obj_t* self);
-uint8_t common_hal_audioio_wavefile_get_channel_count(audioio_wavefile_obj_t* self);
+void common_hal_audioio_rawsample_deinit(audioio_rawsample_obj_t* self);
+bool common_hal_audioio_rawsample_deinited(audioio_rawsample_obj_t* self);
+uint32_t common_hal_audioio_rawsample_get_sample_rate(audioio_rawsample_obj_t* self);
+void common_hal_audioio_rawsample_set_sample_rate(audioio_rawsample_obj_t* self, uint32_t sample_rate);
 
-#endif // MICROPY_INCLUDED_SHARED_BINDINGS_AUDIOIO_WAVEFILE_H
+#endif // MICROPY_INCLUDED_SHARED_BINDINGS_AUDIOIO_RAWSAMPLE_H
