+++ shared-bindings/index.rst
@@ -36,6 +36,7 @@ Module             Supported Ports
 `analogio`         **All Supported**
 `audiobusio`       **SAMD/SAMD Express**
 `audioio`          **SAMD Express**
+`audiocore`        **All with audioio**
 `binascii`         **ESP8266**
 `bitbangio`        **SAMD Express, ESP8266**
 `board`            **All Supported**
